+++ CHANGELOG.md
@@ -15,7 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `starboard-exporter` (app) to v1.2.9.
 - Update `trivy` (app) to v0.18.0.
 - Update `kyverno-policies` (app) to v0.27.0.
-- Update `gel` (app) to v1.0.3.
+
+### Removed
+
+- Remove `gel` (app).
 
 ## [2.3.0] - 2026-08-21