+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `policy-api` (app) to v0.0.10.
 - Update `jiralert` (app) to v0.1.4.
 - Update `starboard-exporter` (app) to v1.2.11.
+- Update `starboard-exporter` (app) to v1.2.15.
 
 ## [2.3.0] - 2026-08-21