+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update `jiralert` (app) to v0.1.4.
 - Update `starboard-exporter` (app) to v1.2.11.
 - Update `starboard-exporter` (app) to v1.2.15.
+- Update `policy-api` (app) to v0.0.12.
 
 ## [2.3.0] - 2026-08-21